requirements.txt
requests==2.31.0
paho-mqtt==1.6.1
python-dotenv==1.0.0
